feat(react-ui-form): add Form createSelectField; migrate AssistantSettings
Add a Form-native createSelectField factory next to SelectField — returns a
FormFieldRenderer (correct prop shape) for fieldMap entries whose choices come
from a runtime list, with an optional sentinel mapping to undefined. Migrate
AssistantSettings to Form (variant='settings'), addressing the nested
modelDefaults.* leaves via dotted JSON-path fieldMap keys. Remove the now-unused
deprecated Settings createSelectField (SettingsSelectField).

Completes the migration of all plugin settings panels off the deprecated
Settings composite; only general-layout consumers remain.
