+++ packages/plugins/plugin-assistant/src/containers/AssistantSettings/AssistantSettings.tsx
@@ -7,7 +7,7 @@ import React, { useMemo } from 'react';
 import { DEFAULT_EDGE_MODELS, DEFAULT_LMSTUDIO_MODELS, DEFAULT_OLLAMA_MODELS } from '@dxos/ai';
 import { type AppSurface } from '@dxos/app-toolkit/ui';
 import { useTranslation } from '@dxos/react-ui';
-import { Settings as SettingsForm, createSelectField } from '@dxos/react-ui-form';
+import { Form, type FormFieldMap, createSelectField } from '@dxos/react-ui-form';
 
 import { meta } from '#meta';
 import { Assistant } from '#types';
@@ -16,28 +16,31 @@ export type AssistantSettingsProps = AppSurface.SettingsProps<Assistant.Settings
 
 export const AssistantSettings = ({ settings, onSettingsChange }: AssistantSettingsProps) => {
   const { t } = useTranslation(meta.profile.key);
-  const fieldMap = useMemo(
+  // Form `fieldMap` is keyed by dotted JSON path, so nested struct leaves are addressed directly.
+  const fieldMap = useMemo<FormFieldMap>(
     () => ({
-      modelDefaults: {
-        edge: createSelectField({ options: DEFAULT_EDGE_MODELS }),
-        ollama: createSelectField({ options: DEFAULT_OLLAMA_MODELS }),
-        lmstudio: createSelectField({ options: DEFAULT_LMSTUDIO_MODELS }),
-      },
+      'modelDefaults.edge': createSelectField({ options: DEFAULT_EDGE_MODELS }),
+      'modelDefaults.ollama': createSelectField({ options: DEFAULT_OLLAMA_MODELS }),
+      'modelDefaults.lmstudio': createSelectField({ options: DEFAULT_LMSTUDIO_MODELS }),
     }),
     [],
   );
 
   return (
-    <SettingsForm.Viewport>
-      <SettingsForm.Section title={t('settings.title', { ns: meta.profile.key })}>
-        <SettingsForm.FieldSet<Assistant.Settings>
-          readonly={!onSettingsChange}
-          schema={Assistant.Settings}
-          fieldMap={fieldMap}
-          values={settings}
-          onValuesChanged={(values) => onSettingsChange?.(() => values)}
-        />
-      </SettingsForm.Section>
-    </SettingsForm.Viewport>
+    <Form.Root
+      variant='settings'
+      schema={Assistant.Settings}
+      readonly={!onSettingsChange}
+      values={settings}
+      onValuesChanged={(values) => onSettingsChange?.(() => values)}
+    >
+      <Form.Viewport scroll>
+        <Form.Content>
+          <Form.Section title={t('settings.title', { ns: meta.profile.key })}>
+            <Form.FieldSet fieldMap={fieldMap} />
+          </Form.Section>
+        </Form.Content>
+      </Form.Viewport>
+    </Form.Root>
   );
 };
