+++ packages/ui/react-ui-form/src/components/Settings/index.ts
@@ -4,4 +4,3 @@
 
 export * from './Settings';
 export { type SettingsFieldMap, type SettingsFieldProps, type SettingsFieldSetProps } from './SettingsFieldSet';
-export { createSelectField, type CreateSelectFieldOptions, type SelectFieldOption } from './SettingsSelectField';
